drug-discovery-prioritisation/requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Cheminformatics
rdkit>=2023.9.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Protein embeddings
transformers>=4.35.0
torch>=2.0.0

# Data acquisition
chembl-webresource-client>=0.10.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0

# PDF generation
reportlab>=4.0.0

# Development
pytest>=7.4.0
black>=23.0.0
